backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videoprocess"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "python-multipart",
    "redis[hiredis]",
    "minio",
    "aiofiles",
    "httpx",
]

[project.optional-dependencies]
api = []
worker = [
    "faster-whisper",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "yt-dlp",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "worker*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
